+++ R/theme-defaults.r
@@ -188,6 +188,11 @@ theme_grey <- function(base_size = 11, base_family = "",
                            hjust = 1, vjust = 1,
                            margin = margin(t = half_line)
                          ),
+    plot.tag =           element_text(
+                           size = rel(1.2),
+                           hjust = 0.5, vjust = 0.5
+                         ),
+    plot.tag.position =  'topleft',
     plot.margin =        margin(half_line, half_line, half_line, half_line),
 
     complete = TRUE
@@ -449,6 +454,11 @@ theme_void <- function(base_size = 11, base_family = "",
                            hjust = 1, vjust = 1,
                            margin = margin(t = half_line)
                          ),
+    plot.tag =           element_text(
+                           size = rel(1.2),
+                           hjust = 0.5, vjust = 0.5
+                         ),
+    plot.tag.position =  'topleft',
 
     complete = TRUE
   )
@@ -566,6 +576,11 @@ theme_test <- function(base_size = 11, base_family = "",
                            hjust = 1, vjust = 1,
                            margin = margin(t = half_line)
                          ),
+    plot.tag =           element_text(
+                           size = rel(1.2),
+                           hjust = 0.5, vjust = 0.5
+                         ),
+    plot.tag.position =  'topleft',
     plot.margin =        margin(half_line, half_line, half_line, half_line),
 
     complete = TRUE