+++ tests/testthat/test-guides.R
@@ -68,13 +68,19 @@ test_that("axis guides are drawn correctly", {
 })
 
 test_that("guides are positioned correctly", {
-  p1 <- ggplot(mtcars, aes(mpg, disp, colour = cyl)) +
+  df <- data.frame(x = 1, y = 1, z = factor("a"))
+
+  p1 <- ggplot(df, aes(x, y, colour = z)) +
     geom_point() +
     labs(title = "title of plot") +
     theme_test() +
-    theme(axis.text.x = element_text(angle = 90, vjust = 0.5)) +
-    scale_x_continuous(breaks = mean(mtcars$mpg), labels = "very very long long axis label") +
-    scale_y_continuous(breaks = mean(mtcars$disp), labels = "very very long long axis label")
+    theme(
+      axis.text.x = element_text(angle = 90, vjust = 0.5),
+      legend.background = element_rect(fill = "grey90"),
+      legend.key = element_rect(fill = "grey90")
+    ) +
+    scale_x_continuous(breaks = 1, labels = "very long axis label") +
+    scale_y_continuous(breaks = 1, labels = "very long axis label")
 
   expect_doppelganger("legend on left",
     p1 + theme(legend.position = "left")
@@ -89,28 +95,28 @@ test_that("guides are positioned correctly", {
     p1 + theme(legend.position = "top")
   )
   expect_doppelganger("facet_grid, legend on left",
-    p1 + facet_grid(am~vs) + theme(legend.position = "left")
+    p1 + facet_grid(x~y) + theme(legend.position = "left")
   )
   expect_doppelganger("facet_grid, legend on bottom",
-    p1 + facet_grid(am~vs) + theme(legend.position = "bottom")
+    p1 + facet_grid(x~y) + theme(legend.position = "bottom")
   )
   expect_doppelganger("facet_grid, legend on right",
-    p1 + facet_grid(am~vs) + theme(legend.position = "right")
+    p1 + facet_grid(x~y) + theme(legend.position = "right")
   )
   expect_doppelganger("facet_grid, legend on top",
-    p1 + facet_grid(am~vs) + theme(legend.position = "top")
+    p1 + facet_grid(x~y) + theme(legend.position = "top")
   )
   expect_doppelganger("facet_wrap, legend on left",
-    p1 + facet_wrap(am~vs) + theme(legend.position = "left")
+    p1 + facet_wrap(~ x) + theme(legend.position = "left")
   )
   expect_doppelganger("facet_wrap, legend on bottom",
-    p1 + facet_wrap(am~vs) + theme(legend.position = "bottom")
+    p1 + facet_wrap(~ x) + theme(legend.position = "bottom")
   )
   expect_doppelganger("facet_wrap, legend on right",
-    p1 + facet_wrap(am~vs) + theme(legend.position = "right")
+    p1 + facet_wrap(~ x) + theme(legend.position = "right")
   )
   expect_doppelganger("facet_wrap, legend on top",
-    p1 + facet_wrap(am~vs) + theme(legend.position = "top")
+    p1 + facet_wrap(~ x) + theme(legend.position = "top")
   )
 
   # padding
@@ -136,4 +142,66 @@ test_that("guides are positioned correctly", {
   expect_doppelganger("legend inside plot, bottom left of legend at center",
     p2 + theme(legend.justification = c(0,0), legend.position = c(.5,.5))
   )
+})
+
+test_that("guides title and text are positioned correctly", {
+  df <- data.frame(x = 1:3, y = 1:3)
+  p <- ggplot(df, aes(x, y, color = factor(x), fill = y)) +
+    geom_point(shape = 21) +
+    # setting the order explicitly removes the risk for failed doppelgangers
+    # due to legends switching order
+    guides(color = guide_legend(order = 2),
+           fill = guide_colorbar(order = 1)) +
+    theme_test()
+
+  expect_doppelganger("multi-line guide title works",
+    p +
+      scale_color_discrete(name = "the\ndiscrete\ncolorscale") +
+      scale_fill_continuous(name = "the\ncontinuous\ncolorscale")
+  )
+  expect_doppelganger("vertical gap of 1cm between guide title and guide",
+    p + theme(legend.spacing.y = grid::unit(1, "cm"))
+  )
+  expect_doppelganger("horizontal gap of 1cm between guide and guide text",
+    p + theme(legend.spacing.x = grid::unit(1, "cm"))
+  )
+
+  # now test label positioning, alignment, etc
+  df <- data.frame(x = c(1, 10, 100))
+  p <- ggplot(df, aes(x, x, color = x, size = x)) +
+    geom_point() +
+    # setting the order explicitly removes the risk for failed doppelgangers
+    # due to legends switching order
+    guides(shape = guide_legend(order = 1),
+           color = guide_colorbar(order = 2)) +
+    theme_test()
+
+  expect_doppelganger("guide title and text positioning and alignment via themes",
+    p + theme(
+      legend.title = element_text(hjust = 0.5, margin = margin(t = 30)),
+      legend.text = element_text(hjust = 1, margin = margin(l = 5, t = 10, b = 10))
+    )
+  )
+
+})
+
+test_that("colorbar can be styled", {
+  df <- data.frame(x <- c(0, 1, 2))
+  p <- ggplot(df, aes(x, x, color = x)) + geom_point()
+
+  expect_doppelganger("white-to-red gradient colorbar, white tick marks, no frame",
+    p + scale_color_gradient(low = 'white', high = 'red')
+  )
+
+  expect_doppelganger("white-to-red gradient colorbar, thick black tick marks, green frame",
+    p + scale_color_gradient(
+          low = 'white', high = 'red',
+          guide = guide_colorbar(
+            frame.colour = "green",
+            frame.linewidth = 1.5,
+            ticks.colour = "black",
+            ticks.linewidth = 2.5
+            )
+        )
+    )
 })